requirements.txt
streamlit==1.40.1
psycopg2-binary==2.9.9
bcrypt==4.2.1
supabase